// apps/mobile/src/navigation/types.ts
export type RootStackParamList = {
  Auth: undefined;
  Pending: undefined;
  MainTabs: undefined;
  ChatThread: {id: string; title: string};
  Fees: undefined;
  Attendance: undefined;
  Report: undefined;
  Progress: undefined;
  Circulars: undefined;
  Notifications: undefined;
  Bus: undefined;
  Profile: undefined;
  ChatsTab: undefined;
  HomeworkTab: undefined;
};

export type MainTabParamList = {
  Home: undefined;
  ChatsTab: undefined;
  HomeworkTab: undefined;
  More: undefined;
};